+++ src/models/file_properties/properties_overwrite.rs
@@ -1,4 +1,4 @@
-use super::PathWithPropertyGroupsArgs;
+use super::AddPropertiesArg;
 
 use crate::{
     anyhow::Result,
@@ -15,7 +15,7 @@ use std::{future::Future, pin::Pin};
 /// Type aliases for readability
 type Request<'a> = PropertiesOverwriteRequest<'a>;
 type Response = PropertiesOverwriteResponse;
-type RequestPayload = PathWithPropertyGroupsArgs;
+type RequestPayload = AddPropertiesArg;
 type ResponsePayload = ();
 
 /// Add properties struct for file request