+++ src/models/file_properties/properties_search.rs
@@ -1,4 +1,4 @@
-use super::{MatchesWithPropertyGroupsResult, QueriesWithTemplateFilterArgs};
+use super::{PropertiesSearchResult, PropertiesSearchArg};
 
 use crate::{
     anyhow::Result,
@@ -15,8 +15,8 @@ use std::{future::Future, pin::Pin};
 /// Type aliases for readability
 type Request<'a> = PropertiesSearchRequest<'a>;
 type Response = PropertiesSearchResponse;
-type RequestPayload = QueriesWithTemplateFilterArgs;
-type ResponsePayload = MatchesWithPropertyGroupsResult;
+type RequestPayload = PropertiesSearchArg;
+type ResponsePayload = PropertiesSearchResult;
 
 /// Add properties struct for file request
 /// https://www.dropbox.com/developers/documentation/http/documentation#file_properties-properties-search