+++ src/models/file_properties/templates_add_for_user.rs
@@ -1,4 +1,4 @@
-use super::{PropertyTemplateArgs, TemplateIdResult};
+use super::{AddTemplateArg, AddTemplateResult};
 
 use crate::{
     anyhow::Result,
@@ -15,8 +15,8 @@ use std::{future::Future, pin::Pin};
 /// Type aliases for readability
 type Request<'a> = TemplatesAddForUserRequest<'a>;
 type Response = TemplatesAddForUserResponse;
-type RequestPayload = PropertyTemplateArgs;
-type ResponsePayload = TemplateIdResult;
+type RequestPayload = AddTemplateArg;
+type ResponsePayload = AddTemplateResult;
 
 /// Add properties struct for file request
 /// https://www.dropbox.com/developers/documentation/http/documentation#file_properties-templates-add_for_user