+++ src/models/file_properties/templates_get_for_user.rs
@@ -1,4 +1,4 @@
-use super::{PropertyTemplateWithTaggedTypeResult, TemplateIdArgs};
+use super::{GetTemplateResult, GetTemplateArg};
 
 use crate::{
     anyhow::Result,
@@ -15,8 +15,8 @@ use std::{future::Future, pin::Pin};
 /// Type aliases for readability
 type Request<'a> = TemplatesGetForUserRequest<'a>;
 type Response = TemplatesGetForUserResponse;
-type RequestPayload = TemplateIdArgs;
-type ResponsePayload = PropertyTemplateWithTaggedTypeResult;
+type RequestPayload = GetTemplateArg;
+type ResponsePayload = GetTemplateResult;
 
 /// Add properties struct for file request
 /// https://www.dropbox.com/developers/documentation/http/documentation#file_properties-templates-get_for_user