+++ src/models/file_properties/templates_remove_for_user.rs
@@ -1,4 +1,4 @@
-use super::TemplateIdArgs;
+use super::RemoveTemplateArg;
 
 use crate::{
     anyhow::Result,
@@ -15,7 +15,7 @@ use std::{future::Future, pin::Pin};
 /// Type aliases for readability
 type Request<'a> = TemplatesRemoveForUserRequest<'a>;
 type Response = TemplatesRemoveForUserResponse;
-type RequestPayload = TemplateIdArgs;
+type RequestPayload = RemoveTemplateArg;
 type ResponsePayload = ();
 
 /// Add properties struct for file request